Refactor, minor perf improvements
